requirements/requirements.txt
Django==1.9.1
markdown==2.6.2
djangorestframework==3.3.2
psycopg2==2.6.1
